+++ app/ui/Login/LogOut.tsx
@@ -1,11 +1,9 @@
 "use client";
 import React from "react";
 import { HiLogout } from "react-icons/hi";
-import { redirect, useRouter } from "next/navigation";
 import { signOut } from "next-auth/react";
 
 const LogOut = () => {
-  const router = useRouter();
   const handleSignOut = async () => {
     await signOut({ callbackUrl: "/" });
   };
@@ -14,7 +12,7 @@ const LogOut = () => {
       className="flex w-full gap-2 items-center text-base font-normal"
       onClick={handleSignOut}
     >
-        <HiLogout size={30} />
+      <HiLogout size={30} />
       Log Out
     </div>
   );
